# `fs.Dir.deleteTree`: Optimize for non-deeply-nested directories