@ifreund 2022-10-05
Very nice work on this PR, I think this is the right direction for the standard library to take!