@squeek502 2022-10-06
A summary of the recent changes and some explanations:

- I reduced the `Iterator` buf size to 1024 for every platform.
  - On Linux, this means that `@sizeOf(std.BoundedArray(StackItem, 16))` is 17288 or ~16.9KiB. Still unsure if that's too large or what a reasonable size to shoot for would be.
  - On Linux this doesn't seem to have much of an impact (see https://github.com/ziglang/zig/pull/13073#discussion_r988415719).
  - On Windows, the larger buf size is more performant, not totally sure what tradeoff should be made there:
    - 'walk-8192.exe' ran
    1.02 ± 0.01 times faster than 'walk-2048.exe'
    1.03 ± 0.01 times faster than 'walk-1536.exe'
    1.07 ± 0.01 times faster than 'walk-1024.exe'
- I renamed `deleteTreeFallback` to `deleteTreeMinStackSize` and made it pub again as per @matu3ba's comment. Let me know if there's a better name for it.
- I fixed the new `deleteTree` implementation failing on Windows (with FileBusy) since I was closing the directory handle after deleting the directory which is a no-go on Windows. I applied the Windows fix to all platforms which introduces some unnecessary complexity on platforms that don't need to close the handle before deletion. I'm unsure about this choice.